+++ NEWS.md
@@ -7,6 +7,7 @@
 * Add `tar_igraph()` and recommend it along with `igraph::find_cycle()` for debugging (#1562, @tylermorganwall).
 * Avoid the unclean shutdown message in the `clustermq` multi-process scheduler by calling `cleanup()` repeatedly until it returns `TRUE`. Uses exponential backoff to avoid excessive CPU load.
 * Use sequential controller in `covr`.
+* Ensure hash stability of vector slices in R >= 4.6.0 (#1566).
 
 # targets 1.11.4